samples: code_relocation_nocopy: Update configuration for nRF54H20 DK
It is no longer needed to specify the read command parameters in dts,
as the flash driver is capable of obtaining those from SFDP structures,
so the redundant properties are removed.

After commit 612fd940fa307ea8a83b97571505d6201ab28d85 got merged, both
the EXMIF and flash nodes in dts are not enabled by default so this
needs to be done in the overlay.

Although nrfutil is still not capable of programming the external
flash on nRF54H20 DK, thus `west flash` will not work in this case,
`nrf54h20dk/nrf54h20/cpuapp` is added as an allowed platform so that
the sample is at least built in CI for it.

(cherry picked from commit 10740d63a54d81fa35a65d90b0e773521174ea69)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: 10740d63a54d81fa35a65d90b0e773521174ea69
Cr-Build-Id: 8704117406820186449
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704117406820186449
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ia2f54942539d2f8ba5037271d923aa626f1a7829
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6936017
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>