net: tcp: Preallocate TX packet for TCP context
TCP context cannot operate w/o a TX packet for buffering transmitted
data. So far this net_pkt was allocated at runtime from the common
packet pool, but this created some not-obvious memory requirement on TCP
and could lead to TX packet starvation in case many TCP connections are
open in parallel. Therefore, allocate this packet structure statically,
as a part of the TCP context instead.

This increases the memory requirement of the TCP context by ~64 bytes,
however if that's a concern for the application, the maximum number of
TX packets can be lowered instead. In return, we get a clear separation
between the number of TCP connections opened, and the amount of packets
that can be transmitted.

(cherry picked from commit eda76ad21b1748725c6d590204ee25d88caa9a82)

Original-Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: eda76ad21b1748725c6d590204ee25d88caa9a82
Cr-Build-Id: 8702939646772595441
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8702939646772595441
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0484cd455a39c65c4fd40d27cb1605018bbf2a52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6973619
Tested-by: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>