soc: nordic: uicr: Fix dependency issue
Although not reproducible locally, it has been observed in CI that the
uicr image will not always be the last image to be run.

To ensure it is the last image to be run we have it depend on the
'image' image when defined.

The uicr image is generated based on all other images in the build and
must therefore run last.

(cherry picked from commit 32c67762568c1aeb4484c2abf621f3f11f0ccf78)

Original-Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
GitOrigin-RevId: 32c67762568c1aeb4484c2abf621f3f11f0ccf78
Cr-Build-Id: 8703460579189758961
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8703460579189758961
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0875ad3c574e7b4002fc42eb92febe6ff655ddec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6952511
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>