scripts: west_commands: runners: openocd path fix
There is a fix that is needed for building on Linux and
running on Windows.  Without this fix, the scripts throws
errors and fails because of the '\' instead of '/'.  For
portability, the script should not use literal slashes.

(cherry picked from commit e20c04a32094684f242a5373d1cbf867473ce3c3)

Original-Signed-off-by: Bill Waters <bill.waters@infineon.com>
GitOrigin-RevId: e20c04a32094684f242a5373d1cbf867473ce3c3
Cr-Build-Id: 8703460579189758961
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8703460579189758961
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ib452d06abecd3b63d82697e840f1aa0af3dcf19a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6954581
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>