posix: net: move gethostname() from static inline to normal function
Convert `gethostname()` from a static inline wrapper to a normal function
with regular linkage.

This change was prompted mainly as housekeeping for #95811, but a
secondary reason was to make this a linkable symbol so that the API works
at a binary level as well (which could be provided by the toolchain as
well).

(cherry picked from commit f071df29a745f2304c80658a06bea7d66a71ac43)

Original-Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>
GitOrigin-RevId: f071df29a745f2304c80658a06bea7d66a71ac43
Cr-Build-Id: 8702939646772595441
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8702939646772595441
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ifb1e97d97d0e416ca3f411eee9fb60d097cca127
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6974170
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>