samples: boards: nxp: add new sample for frdm_mcxa156
Add new sample lpdac_opamp_lpadc for frdm_mcxa156.

In this example:
1. The LPDAC outputs a 300mv voltage to the
non-inverting terminal of the OPAMP.

2. The OPAMP operates in the non-inverting amplifier
mode, the inverting terminal of the OPAMP is grounded.

3. The channel 0 of the LPADC is connected to the output
of the OPAMP, and the channel 1 of the LPADC is connected
to the output of the DAC (the non-inverting input terminal
of the OPAMP).

4. The ideal sampling range of OPAMP output channel is
specified by 'ideal-sample-range', and 'nxp,opamps' are
used to specify LPADC samples which OPAMP's output and
which ADC channel is connected to the OPAMP output.

5. The OPAMP non-inverting input is 300mV, and the OPAMP
gain is initialized to 1. Therefore, the first round OPAMP
output (~300mV) will not reach the minimum value of the
'ideal-sampling-range'. Therefore, the ADC driver will
increase the OPAMP gain through 'opamp_set_gain'. The second
round OPAMP output (~900mV) is still below the minimum value,
so the gain will continue to increase. The third round OPAMP
output (~1500mV) is already in the ideal range, so the gain
will not be adjusted later.

(cherry picked from commit 857347a2bb880e884bb74dda0baf9e7896284334)

Original-Signed-off-by: Zhaoxiang Jin <Zhaoxiang.Jin_1@nxp.com>
GitOrigin-RevId: 857347a2bb880e884bb74dda0baf9e7896284334
Cr-Build-Id: 8702803750045255809
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8702803750045255809
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I13668807dc489353490fd27c52b381348ff21c65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6982017
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>