mqtt_sn: process_ping: fix infinite loop
When next_ping == now, the code scheduled the workqueue for the current
time. On native_sim, this meant that the system workqueue thread was stuck
in an infinite loop because it kept processing the MQTT work over and over
again and the current timestamp could not advance anymore.

I didn't investigate why the yield inside the workqueue didn't help or why
native_sim can't advance time when one of the threads is stuck, but
changing the condition to >= inside mqtt_sn solves this issue.

I discovered this while running zephyr.exe through strace for up to 60
minutes. I guess that the performance overhead makes it more likely for the
workqueue handler to be run while next_ping == now, but I didn't verify
that, because it takes a long time to trigger the bug.

(cherry picked from commit 5bba1edb6e64ccc309612f6039c08a31f2cd7646)

Original-Signed-off-by: Michael Zimmermann <michael.zimmermann@sevenlab.de>
GitOrigin-RevId: 5bba1edb6e64ccc309612f6039c08a31f2cd7646
Cr-Build-Id: 8703437929377721809
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8703437929377721809
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I70b2e6033bb56fea0283d70a8a8c9f6e39c1a73a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6963383
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>