drivers: serial: silabs: Update ISR to check for DMA device presence
Since the symbol CONFIG_UART_ASYNC_API is shared between EUSART and USART
drivers, it creates a scenario where EUSART can have
CONFIG_UART_SILABS_USART_ASYNC=y but with no DMA device declared for it
in the DTS (because we only want async API for the USART driver).
We handled this case by disabling async transfer when the DMA
device is null, but when we activate CONFIG_PM, we have a hard fault due
to null pointer dereference in the ISR.
This bug was only discovered today when trying to enable CONFIG_PM over the
uart_async_api test, which led to a hard fault in the ISR. This occurs
because we only enable interrupts when the PM is enabled in the "poll_out"
function to put the PM lock back.

(cherry picked from commit c37ffd396b02b61b0b9744f5d733bfc4a3e32e14)

Original-Signed-off-by: Martin Hoff <martin.hoff@silabs.com>
GitOrigin-RevId: c37ffd396b02b61b0b9744f5d733bfc4a3e32e14
Cr-Build-Id: 8703392631739715201
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8703392631739715201
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I606c26be47626ed18873d1e5ab20321de769cb88
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6965599
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>