boards: rt1180: fix lpspi with dma
Since this board is by default having hyperram as SRAM, need to set
NOCACHE_MEMORY when using dma to make sure the memory for the tcd pools
is not cached.

The previous approach to set zephyr,sram to dtcm just
caused an error because zephyr,dtcm was also set to dtcm, and the
default logic of the driver is put the tcdpools in dtcm, so what
happened was these buffer got linked over the rest of the app causing
massive failing of the image to work at all.

So instead of doing that we need to just keep running from hyperram but
set CONFIG_NOCACHE_MEMORY to have a place to put the tcd pools that wont
be cached.

(cherry picked from commit 11d24034e0786ef9596b7ef9ac8b56797d5d19d6)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 11d24034e0786ef9596b7ef9ac8b56797d5d19d6
Cr-Build-Id: 8703347329197623089
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8703347329197623089
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8af51f54f3246a9904622802fbefa2c3744c3036
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6968019
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>