net: if: Fix assert when checking lladdr existence
As the linkaddr->addr is no longer a pointer, the original
assert check is not working as it should. So use the length
of the linkaddr to verify that it is > 0.

(cherry picked from commit 221aca4bb9217e666d44f88c0042d5bd426b75a2)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: 221aca4bb9217e666d44f88c0042d5bd426b75a2
Cr-Build-Id: 8703437929377721809
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8703437929377721809
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8da796b10ccbdf8dd5971e44d80e6832b92020bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6963411
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>