scripts: get_maintainer: support file groups
This new section allows defining a group of files in an area and makes
it possible to assign collaborators to the file group being defined.

The purpose of this new section is to allow fine tuning who is added as
reviewer when files change in a group. It is especially useful in large
areas with hundreds of files, for example platform areas.

(cherry picked from commit 9165fb4927a1b77ee3fc0eb09539dbdde6098280)

Original-Signed-off-by: Anas Nashif <anas.nashif@intel.com>
GitOrigin-RevId: 9165fb4927a1b77ee3fc0eb09539dbdde6098280
Cr-Build-Id: 8697820917959855649
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8697820917959855649
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I662ccab5d7590fc4aa6e3da3146c5e9ad896395c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7171022
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>