pm: refactor PM_S2RAM_CUSTOM_MARKING option to be promptless
The config PM_S2RAM_CUSTOM_MARKING is not an optional config for a
user to select, it is required by some soc implementations of S2RAM,
in which case it must be selected by the soc.

Refactor the configuration to be HAS_PM_S2RAM_CUSTOM_MARKING, and
make the currently only soc which needs it select it. Then update
samples which previously had to select this option for this soc.

(cherry picked from commit 1767f131aac706f978eb93bd548b7e7419a0457d)

Original-Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>
GitOrigin-RevId: 1767f131aac706f978eb93bd548b7e7419a0457d
Cr-Build-Id: 8702916994902323137
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8702916994902323137
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Id9cfd7b68d20b289179faa52f7543d9033e777ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6977257
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>