Bluetooth: BAP: Add bt_bap_ep_get_conn
Add a new function, bt_bap_ep_get_conn, which returns
the ACL connection for the endpoint. This works because
endpoints are specific to an ACL in BAP. The function
returns a pointer with a new reference similar to the
lookup functions from conn.h

The conn pointer was not added to the bt_bap_ep_info
struct, as doing so would be more likely to cause
reference leaks if the caller did not care about the
conn pointer when using bt_bap_ep_get_info. It can be
added later if that is requested.

(cherry picked from commit 59a1c6e0de7716b947b892bc6213ccf591b38658)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 59a1c6e0de7716b947b892bc6213ccf591b38658
Cr-Build-Id: 8693426963900677041
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8693426963900677041
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I58cdd2ab317c04ff122e71a0f6fc87575db0dd40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7402172
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>