boards: st: n6: avoid access to last 32B of memory-mapped XSPI device
System might become unstable when the last 32 bytes of a XSPI attached
are accessed in memory-mapped mode with cache enabled.
In order to prevent this, remove the last 32-bytes of the PSRAM on the
stm32n6570_dk as well as NOR on the nucleo n657x0_q.
Partitions should also be defined in a way to avoid accessing to the
last 32-bytes of the device.

Issue is described in the STM32N6xxxx errata sheet ES0620 Rev 1:
2.4.5 Deadlock or data corruption when DQS is enabled and the wrap
request includes the last address of the memory

(cherry picked from commit 8675d738e79500160b13ec65d99030ba82bd8c44)

Original-Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
GitOrigin-RevId: 8675d738e79500160b13ec65d99030ba82bd8c44
Cr-Build-Id: 8702667854123090769
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8702667854123090769
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I783c47bd035536be466841389c58bee474f0f879
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6988846
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>