Bluetooth: Controller: Deprecate bt_ctlr_set_public_addr
Deprecate the bt_ctlr_set_public_addr function.
It is unused in the tree, and there is a vendor specific HCI
command that can be used for the same purpose.
The function does not have any tests and is basically
unmaintained.
Additionally it is the only function in the public
controller API, and removing it would allow us to
remove that part completely as unused.

(cherry picked from commit b07be641e788d6be277ccfe8931149c993f9574e)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: b07be641e788d6be277ccfe8931149c993f9574e
Cr-Build-Id: 8701014460813377745
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701014460813377745
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Id54ac7a2a93839a5fde75f3bf8ff90dce744ceab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7033073
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>