shell: history rewrite to use k_heap instead of ring_buffer
The ring_buffer API was not ideal for this use case, as it is designed for
concurrent access, which is not needed here. The new implementation using
k_heap is more readable and easier to maintain.

There is a slight penalty in memory usage & effectiveness due to the
overhead of k_heap, but since this isn't a production-critical path, the
trade-off is acceptable.

(cherry picked from commit 683819f862020aa1ca41d7807a5020c8c2fa5ad3)

Original-Signed-off-by: Måns Ansgariusson <Mansgariusson@gmail.com>
GitOrigin-RevId: 683819f862020aa1ca41d7807a5020c8c2fa5ad3
Cr-Build-Id: 8701014460813377745
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701014460813377745
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7df2a6bfb9b5b6cd76a44921757d327e877c7871
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7040816
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>