tests: Bluetooth: ASCS: Modify unittests to native_sim
Modify the unit tests to use the native_sim board instead of the
unit_testing board.
The native_sim board allows us to use the kernel in a more
regular way, while still being able to treat the tests as
unittest.

To support mocking we cannot rely on the default
Kconfig options and values. The required Kconfig options
to satisfy dependencies are overridden in the Kconfig
file and the necessary Kconfig files are sourced.
We cannot enable e.g. CONFIG_BT as that would attempt
to pull in a lot of files we do not want in this test.

This commit also removes unused includes in the CMakelists,
and cleans up the code.
Since the kernel is actually running, we also need to support
that e.g. k_work items are actually scheduled, and that we have
to wait for them to finish. test_drain_syswq has been implemented
as a way to drain the system workqueue before we check for
callbacks or perform subsequent actions.

(cherry picked from commit 50a5cb3ec680b8021f1800ae2653849a0f794c66)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 50a5cb3ec680b8021f1800ae2653849a0f794c66
Cr-Build-Id: 8696597854213163953
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696597854213163953
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I52b86da2a3fafd20986758c3be192ef8f71bf022
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7215698
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>