tests: Bluetooth: CAP: Modify commander unittests to native_sim
Modify the unit tests to use the native_sim board instead of the
unit_testing board.
The native_sim board allows us to use the kernel in a more
regular way, while still being able to treat the tests as
unittest.

To support mocking we cannot rely on the default
Kconfig options and values. The required Kconfig options
to satisfy dependencies are overridden in the Kconfig
file and the necessary Kconfig files are sourced.
We cannot enable e.g. CONFIG_BT as that would attempt
to pull in a lot of files we do not want in this test.

This commit also removes unused includes in the CMakelists

(cherry picked from commit 90b05da76a2eff955b2887a90331dee8695afcdd)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 90b05da76a2eff955b2887a90331dee8695afcdd
Cr-Build-Id: 8696597854213163953
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696597854213163953
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I01708927c5498c4d470e93bf155dcdbc70f17675
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7215699
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>