tests: Bluetooth: CAP: Modify initiator unittests to native_sim
Modify the unit tests to use the native_sim board instead of the
unit_testing board.
The native_sim board allows us to use the kernel in a more
regular way, while still being able to treat the tests as
unittest.

To support mocking we cannot rely on the default
Kconfig options and values. The required Kconfig options
to satisfy dependencies are overridden in the Kconfig
file and the necessary Kconfig files are sourced.
We cannot enable e.g. CONFIG_BT as that would attempt
to pull in a lot of files we do not want in this test.

This commit also removes unused includes in the CMakelists

(cherry picked from commit bfe66fc548411e0300bb05452e52d94a907110fe)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: bfe66fc548411e0300bb05452e52d94a907110fe
Cr-Build-Id: 8696665805680353105
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8696665805680353105
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I741c74770355de2ad946b726a6b8d33149fda0d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7211282
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>