drivers can_native_linux: Set socket to be closed on exec
If the process does an exec() (or fork, or..) all descriptors are kept
open by default, unless O_CLOEXEC is set for them.
(This default behaviour is usefull for stdin/out/err so that new process
is kept connected to them, but it is very rare for it to be usefull for
any other descriptor)

In general this leads to descriptors being kept open unnecessarily,
which either will block other process from using them (for example
if the child survives the parent but it does something else while keeping
a port open).
Or for a "leak" which unnecessarily uses descriptors and memory in the
child process.

Let's ensure we do not leak it for this component as we do not need it.

(cherry picked from commit 01118e26570ca9358f96800276f83c48f4a5a6b0)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 01118e26570ca9358f96800276f83c48f4a5a6b0
Cr-Build-Id: 8701988378913007633
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701988378913007633
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If79c65313391b386322873e1d43d5e52c24dde6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7009355
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>