drivers eeprom_simulator_native: Set EEPROM file to close on exec
If the process does an exec() (or fork, or..) all descriptors are kept
open by default, unless O_CLOEXEC is set when opening them.
This is usefull for stdin/out/err so that new process is connected to
them, but it is very rare for it to be usefull for any other descriptor.

In general this leads to descriptors being kept open unnecessarily,
which either will block other process from getting them (for example
if the child survives the parent but it does something else).
Or for a "leak" which unnecessarily uses descriptors and memory in the
child process.

Let's ensure we do not leak it for this component as we do not need it.

(cherry picked from commit d91565cd3aa51a1781bef3a40cdbb5387cb356f3)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: d91565cd3aa51a1781bef3a40cdbb5387cb356f3
Cr-Build-Id: 8701988378913007633
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701988378913007633
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8e9925eb310afbc75b6aaf0e654accaa6eedce80
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7009356
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>