drivers/net/nsos: Set socket to be closed on exec
If the process does an exec() (or fork, or..) all descriptors are kept
open by default, unless O_CLOEXEC is set when opening them.
This is usefull for stdin/out/err so that new process is connected to
them, but it is very rare for it to be usefull for any other descriptor.

In general this leads to descriptors being kept open unnecessarily,
which either will block other process from getting them (for example
if the child survives the parent but it does something else).
Or for a "leak" which unnecessarily uses descriptors and memory in the
child process.

Let's ensure we do not leak it for this component as we do not need it.

(cherry picked from commit b850715a2914d5ef87cce2364b0729fdb9442e9b)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: b850715a2914d5ef87cce2364b0729fdb9442e9b
Cr-Build-Id: 8701988378913007633
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701988378913007633
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If44873a96dbe01c9458adebae85a2c9e12b028c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7009360
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>