include: video: add size field to video_format structure
Add size field to the video_format structure which needs to be set by the
driver and exposed to the application.

For uncompressed formats, this is the size of the raw data buffer in
bytes, which could be the whole raw image or a portion of the raw image
in cases the receiver / dma supports it.

For compressed formats, this is the estimated maximum number of bytes
required to hold a complete compressed frame.

(cherry picked from commit b69e1d550a54db51eafcf2fa8ab1e8ee75daeff2)

Original-Signed-off-by: Hugues Fruchet <hugues.fruchet@foss.st.com>
Original-Signed-off-by: Phi Bang Nguyen <phibang.nguyen@nxp.com>
GitOrigin-RevId: b69e1d550a54db51eafcf2fa8ab1e8ee75daeff2
Cr-Build-Id: 8701444795108161617
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701444795108161617
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I14945fcd0d6522d0be3bebb5804c3c64689ac73c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7031182
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>