kernel: dynamic: update storage size for pool of dynamic thread stacks
Commit 5c5e17f introduced a subtle regression when userspace was
configured on architectures requiring guard pages.

Prior to 5c5e17f, the assumption was that guard pages would be included in
`CONFIG_DYNAMIC_THREAD_STACK_SIZE`, and that was something that the caller
of `k_thread_stack_alloc()` would need to be aware of, although it was not
documented at all, unfortunately.

It seems that 5c5e17f intended to remove the need for that assumption, but
the necessary conditions for doing so had not been met.

Update pool storage size to account for guard pages, which ensures that
users can access every byte of `CONFIG_DYNAMIC_THREAD_STACK_SIZE` rather
than needing to be aware that guard pages would be included in the
requested size.

The compromise is a more intuitive API at the cost of more storage space
for the pool of thread stacks when userspace is enabled.

(cherry picked from commit 6c01157fef4385e12ec7d271487c690b23659a0b)

Original-Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>
GitOrigin-RevId: 6c01157fef4385e12ec7d271487c690b23659a0b
Cr-Build-Id: 8702124274349247249
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8702124274349247249
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7273ecf0301efdf5b9e8d81ded1130b39b0f09b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7006406
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>