tests: kernel/smp: intel_adsp needs more relaxing NOPs
It has been observed that, during the switching stress test,
the context switching becomes very slow due to enough CPUs
doing RCW transaction on hardware bus (e.g. spin locks).
Since the number of NOPs are the same for all CPUs, they are
simply entering and exiting the relax loop at the same time,
and hitting the bus with RCW transactions at the same time.
Not exactly a deadlock but it slows down the execution
enough to result in the test timing out. The SoC layer has
added an option to offset the number of NOPs per CPU by
adding the CPU to the number of NOPs. So enabling it for
the Intel ADSP boards to workaround the above mentioned
issue. I haven't encountered another slowdown after turning
on this option. So hopefully this lowers the probability of
that happening such that a simple retry can pass the test.

(cherry picked from commit 579225f37049760a06ba97c965b3646f6ffa594f)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: 579225f37049760a06ba97c965b3646f6ffa594f
Cr-Build-Id: 8700357631318882593
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700357631318882593
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie7e5356dcf548d2d806ec9423e9baf4fcf91bf7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7070472
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>