net: ip: account for the size in the inet_ntop code path
The code was writing to the dst without a verification check on
size which is not appropriate. The guard on the arguements should
be enforced and so just ensure the size is larger then the
definition of the strings from POSIX and return an error in those
cases.

(cherry picked from commit 2bfd76425ae612acd32b8f4f229572005edce1e4)

Original-Signed-off-by: Charles Hardin <ckhardin@gmail.com>
GitOrigin-RevId: 2bfd76425ae612acd32b8f4f229572005edce1e4
Cr-Build-Id: 8701558043672538497
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701558043672538497
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ieb9a1d43c3a33c7869bdaacdf3241b9bd8207526
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7016376
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>