pm: Remove "power state consistency" check
This check is totally wrong, min-residency-us and exit-latency-us are
not related to each other at all except that the end of the residency
period occurs when the exit starts, so they are totally orthogonal
values and should not be checking if one is less than the other or
anything like this. Even in the code for the default policy manager,
they are added together, so this check is incompatible / in disagreement
with that code, so it should be removed.

(cherry picked from commit 74d715c5ef92820fb9e504baa0586c6e4acbbf5d)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 74d715c5ef92820fb9e504baa0586c6e4acbbf5d
Cr-Build-Id: 8701467445742638449
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701467445742638449
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I585ee936f33233aacaff0b069c7e2e84be934c6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7027945
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>