Bluetooth: Host: Change WRN->DBG for not connect for TX
The bt_conn_tx_processor logged connections not in the
connected state as LOG_WRN, but this is a case that is handled
and could occur during regular behavior.
For this reason the log statement was changed to LOG_DBG,
as the developer/user does not need to worry about it.
Also modifies the statement to log the conn->state for
debugging purposes.

(cherry picked from commit d4afa1989fda9c6919f97ecbbbe17a5c0a390032)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: d4afa1989fda9c6919f97ecbbbe17a5c0a390032
Cr-Build-Id: 8701014460813377745
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701014460813377745
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iad540d6af8c65faa7d821a17ff535941279c1b4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7033045
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>