net: dns: extend the service resolver command line for ptr, srv, addr
The address resoution from RFC6763 is generally a PTR, SRV, TXT, and
an A or AAAA. Records, so this change is mainly to start drawing a
clear functional change between the DNS query shell and the DNS
service shell that should be more of a "avahi-browse" or "dns-sd"
functional style.

So, this is not a very robust implementation since it is a chain of
queries when the goal should be to get additional records from the
message itself but those changes can be added iteratively.

Basic point of this change is to show the procedure for the "browse"

- PTR service resolution
- SRV query from the result of the PTR
- AAAA (or A) from the result of the SRV

TXT records are ignored for now but should be added into this as
additional record support is extended.

(cherry picked from commit 7158f33c2b00db27156656f811154b5c8b05c0eb)

Original-Signed-off-by: Charles Hardin <ckhardin@gmail.com>
GitOrigin-RevId: 7158f33c2b00db27156656f811154b5c8b05c0eb
Cr-Build-Id: 8700946513353501409
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700946513353501409
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ib710c01a54d178348e14deea70c982a578a2042e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7042001
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>