samples: video: avoid K_FOREVER wait on initial buffer allocation
On samples startup, video buffers are being allocated in order to
hold the captured data. Currently the timeout is set to K_FOREVER,
leading to having those samples waiting forever if there isn't
enought memory available on the platform.
Considering that this is done at beginning of the sample and at
that moment there isn't reason to having to wait for the
availability of memory, the timeout is replaced by K_NO_WAIT which
will allow to have the function returning with an error if the
allocation is failing.

(cherry picked from commit f2a3f11427d6de78967926ef5b9cdef643e7b335)

Original-Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
GitOrigin-RevId: f2a3f11427d6de78967926ef5b9cdef643e7b335
Cr-Build-Id: 8701014460813377745
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8701014460813377745
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ifdadb80743aa83f6b54684f2469e0098ee47b2c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7040807
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>