drivers: adc: stm32: Rename sequencer and oversampler and fix macro issue
In STM32 ADC binding, rename the possible values of the sequencer and
oversampler properties to use lowercase string, similar to the internal
regulator.

Adapts the driver and the dtsi with the new values.

Fixes a macro issue in the driver. Since the value from the dtsi didn't
start with internal_regulator_, the reconstruction of the defines by
the macro ANY_ADC_INTERNAL_REGULATOR_TYPE_IS was missing this prefix and
the comparison failed. Add a new argument to the IS_EQ_STRING_PROP to be
able to insert such a prefix.

(cherry picked from commit 625e2acda878a2bda389ca6e2a28db96dee35aaa)

Original-Signed-off-by: Guillaume Gautier <guillaume.gautier-ext@st.com>
GitOrigin-RevId: 625e2acda878a2bda389ca6e2a28db96dee35aaa
Cr-Build-Id: 8700833267110927073
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8700833267110927073
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I10558183760002a68254b8cf34e458b146a7b1a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7050398
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>